feat: extract headless core from the web emulator
- add the `boytacean-core` package under `frontends/web/core`, exposing a `GameBoyCore` that carries all of the emulation logic with no React, EmuKit UI or bundler specific dependency
- build the core on top of the new headless `emukit/logic` entry point, so the loop, event and framerate accounting are shared with the web app
- replace the direct `localStorage` usage with a `StorageAdapter` abstraction, defaulting to the Web Storage API and falling back to a null adapter outside the browser
- take the WASM binary path and the extra settings as constructor options, removing the `require` calls from the emulation logic
- move the WASM `window` callbacks into the core, dispatching them to the active core instance instead of `window.emulator`
- move `ts/palettes.ts` into the core package, re-exported from `ts/index.ts` for backwards compatibility
- reduce `GameboyEmulator` to an EmuKit adapter over the core, keeping only the UI bound surface (sections, help, debug panels, ROM metadata and playlist)
- alias `boytacean` to the local WASM output and `emukit/logic` to the EmuKit entry point, for both Parcel and TypeScript resolution
- raise the EmuKit dependency to `^0.10.24`, the first version shipping the headless entry point

Only a single core instance can be active at a time, since the WASM callbacks are bound to the `window` namespace. Per instance callbacks are tracked separately.
