+++ frontends/web/core/index.ts
@@ -1,0 +1,1 @@
+export * from "./ts";
