+++ frontends/web/core/package.json
@@ -1,0 +1,26 @@
+{
+    "name": "boytacean-core",
+    "version": "0.12.1",
+    "description": "The headless core of Boytacean, ready to be embedded",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/joamag/boytacean.git"
+    },
+    "license": "Apache-2.0",
+    "scripts": {
+        "lint": "eslint .",
+        "pretty": "prettier --config .prettierrc \"./**/*.{js,ts,tsx,json}\" --write",
+        "test": "npm run test:mocha",
+        "test:mocha": "mocha --node-option experimental-transform-types --recursive test"
+    },
+    "source": "index.ts",
+    "peerDependencies": {
+        "boytacean": "^0.12.1",
+        "emukit": "^0.10.24"
+    },
+    "devDependencies": {
+        "mocha": "^11.1.0",
+        "prettier": "^3.5.3",
+        "typescript": "^5.8.2"
+    }
+}
