+++ frontends/web/core/ts/index.ts
@@ -1,0 +1,3 @@
+export * from "./gb";
+export * from "./palettes";
+export * from "./storage";
