+++ frontends/web/core/ts/palettes.ts
