+++ frontends/web/package.json
@@ -24,7 +24,7 @@
         "@typescript-eslint/eslint-plugin": "^7.18.0",
         "@typescript-eslint/parser": "^7.18.0",
         "buffer": "^6.0.3",
-        "emukit": "^0.10.22",
+        "emukit": "^0.10.24",
         "eslint": "^8.57.1",
         "eslint-plugin-import": "^2.31.0",
         "eslint-plugin-react-hooks": "^5.2.0",
@@ -42,6 +42,8 @@
         "webgl-plot": "^0.7.2"
     },
     "alias": {
+        "boytacean": "./lib/boytacean.js",
+        "emukit/logic": "./node_modules/emukit/logic.ts",
         "process": {
             "global": "process"
         }
