+++ frontends/web/ts/index.ts
@@ -1,4 +1,4 @@
+export * from "../core";
 export * from "./gb";
-export * from "./palettes";
 export * from "./playlist";
 export * from "./util";
